Avoid crashes if inputs are too far apart
This explicitly limits a cast from float to int in GetNumberOfSteps to avoid undefined behavior.

It also adds a parameter to limit the number of samples that can be requested in a single call to StrokeModeler::Update/Predict (somewhat arbitrarily defaulting that to 100k), so that requesting too many samples in a single call will result in an invalid status instead of a crash.

(That seems a real possibility, given that a client could be suspended mid-stroke and then produce a very delayed kUp event when resumed far later, if it's using some absolute measurement for its input time.)

PiperOrigin-RevId: 438377887