Avoid segfault in StylusStateModeler::Query
This can occur if distance is never within min_distance of the target position. Since min_distance starts at positive infinity, that seems like it could only happen if the modeled position has gotten into NaN-land somehow, but I haven't pinned down the underlying bug.

This uses the same edge-case logic as when the positions_and_states_ queue has a single item (in which case the body of the loop that sets closest_segment_index is never run). So that's moved to cover both edge-cases.

PiperOrigin-RevId: 438859784