Add WobbleSmootherParams.position_resolution
One issue in WobbleSmoother is that input quantization can make the speed look incorrectly high. For example, an epsilon move diagonally across a pixel boundary may get represented in the input as a pixel-width difference in the x and y coordinates, and thus a sqrt(2*position_resolution^2)/dt difference in the speed.

This adjusts the speed perceived by WobbleSmoother down by that amount to ensure that slow motions near pixel boundaries don't cause smoothing to jump on and off.

PiperOrigin-RevId: 450040312