Update stroke_model::PositionModelerParams comments and naming.
Rename spring_mass_constant to spring_constant and use the reciprocal of the current default value to accurately reflect Hooke's Law. The current name should have actually been mass_divided_by_spring_constant.

State that the modeling assumes the mass to be 1. The use of mass in the spring constant and drag constant was inconsistent, they should either both have it or neither and we always assume mass = 1.

PiperOrigin-RevId: 451198595