Weight average of speed in WobbleSmoother
Instead of averaging speed naively across the samples, weight each sample by the duration between that and the previous sample. This eliminates some strange edge-case logic for zero-duration samples, which means that duplicate inputs can't cause you to suddenly switch between smoothed and unsmoothed.

If the duration of the entire interval is zero, use the unsmoothed position to match previous behavior.

PiperOrigin-RevId: 453971150