Add support for saving and restoring state and for disabling the internal predictor.
This consists of adding Save() and Restore() to StrokeModeler and most of the internal modelers it uses. A new type of PredictionParams is added to allow the StrokeModeler to run without allocating a predictor. In case the new save points do get used in combination with an enabled predictor, the predictor types are updated to support copy so that the StrokeModeler can save their state.

PiperOrigin-RevId: 509342102