Don't clear the output vector in StrokeModeler::Update
This is more consistent with how this class is expected to be used. StrokeModeler::Update extends a smoothed stroke with new input, while StrokeModeler::Predict replaces an earlier prediction with a new one.

Also update the usage documentation, which missed being updated when Update and Predict were modified to take an output vector as a parameter instead of returning a newly-constructed vector.

PiperOrigin-RevId: 520683308