Add workspace name to workspace
Setting this is best practice:
https://bazel.build/rules/lib/globals/workspace#workspace

It makes it clearer what name consumers should use to refer to this. This pattern also makes it possible for workspace.bzl to have rules that work consistently for this project and its consumers when they need to refer to file in this project.

Also add the current name and version to MODULE.bazel.

PiperOrigin-RevId: 576282143