Fix angle-based upsampling
Previous logic was confusing estimated velocity at end of span with estimated acceleration.

The previous logic also sometimes estimated no move as a 180-deg turn due to over-estimation of drag. This can still be a problem if you clamp the acceleration due to drag to 100% of the original velocity, since then a small acceleration can seem like it would cause a sharp turn when drag eliminates all/most of the original velocity. Instead, ignore drag when estimating the sharpness of a turn that would be caused by an input pre-upsampling.

Finally, add unit tests for Vec2::AbsoluteAngleTo, which should have been added in the previous change.

PiperOrigin-RevId: 602794646