Stop supporting non-Bzlmod Bazel builds
Bzlmod is the preferred system for Bazel dependency management, and the plan is to phase out the old WORKSPACE dependency setup by 2025:
https://blog.bazel.build/2023/07/24/whats-new-with-bzlmod.html#high-level-roadmap-for-the-future

We're using Bzlmod in our CI builds currently, and development will be easier if we're not trying to maintain two copies of dependency configuration.

Also remove an empty WORKSPACE.bzlmod file. This is processed in addition to MODULES.bazel when Bzlmod is enabled, but an empty one doesn't do anything.

PiperOrigin-RevId: 607689427