Add constants to both the Ink stroke modeler and Ink for kPi and kSqrt2.
They're defined with std::numbers when C++20 is available and <cmath> constants for older C++ versions.

This makes the code portable to Windows as <cmath> constants aren't available in Windows build environments by default.  This only builds for Windows with C++20.  For C++17 support we would need to also add `#define _USE_MATH_DEFINES` before including `<cmath>`.  We can address that when or if we have a C++17 Windows builder running.

PiperOrigin-RevId: 624254804