Add position to StylusState in preparation
This is the first step to move velocity based interpolation into the input modeler. Splitting this out as a first step to keep it a little cleaner since this is just moving a value into the struct and changing the call to Update accordingly but all the avalues / tests are still returning the same thing.

PiperOrigin-RevId: 658058051