Estimate angular traversal more consistently
This is supposed to be what the change in angle of velocity would be if there is no upsampling. Previously, this excluded drag, but that wasn't the correct decision. Discrete simulation of drag can spuriously cause the direction of motion to reverse, which is another reason why upsampling in that situation might be desirable.

Use absl::Status matchers in test.

PiperOrigin-RevId: 798224352