Estimate angular traversal more consistently
This is supposed to be what the change in angle of velocity would be if there is no upsampling. Previously, this excluded drag, but that wasn't the correct decision. When discretely simulating motion on a sharp corner, the turnaround can happen almost entirely on a single simulation step. Simulated drag can affect _which_ step that happens on, so failing to take that into account can mean this upsampling is skipped.

Discrete simulation of drag can spuriously cause the direction of motion to reverse, which is another reason why upsampling in that situation might be desirable.

Use absl::Status matchers in test.

PiperOrigin-RevId: 800127834